refact all P2P DB + add new tables for P2P

+++ Web-Api.online/Models/Tables/P2PSellerPaymentTableModel.cs
@@ -3,5 +3,5 @@ namespace Web_Api.online.Models.Tables;
 public class P2PSellerPaymentTableModel
 {
     public long P2PSellerId { get; set; }
-    public long PaymentId { get; set; }
+    public int PaymentId { get; set; }
 }
