+++ Web-Api.online/Models/Tables/P2PUserTableModel.cs
@@ -5,9 +5,9 @@ public class P2PUserTableModel
     public long Id { get; set; }
     public string UserId { get; set; }
     public decimal Price { get; set; }
-    public long FiatId { get; set; }
+    public int FiatId { get; set; }
     public decimal LimitFrom { get; set; }
     public decimal LimitTo { get; set; }
     public decimal Available { get; set; }
-    public long CryptId { get; set; }
+    public int CryptId { get; set; }
 }
